Export Report

+++ Course-Signup-System/Controllers/ClassController.cs
@@ -1,6 +1,7 @@
 ﻿using Course_Signup_System.DTO;
 using Course_Signup_System.Repositories;
 using Course_Signup_System.Services;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 
@@ -8,6 +9,7 @@ namespace Course_Signup_System.Controllers
 {
     [Route("api/[controller]")]
     [ApiController]
+    [Authorize]
     public class ClassController : ControllerBase
     {
         private readonly IClassService _classService;
@@ -20,8 +22,20 @@ namespace Course_Signup_System.Controllers
         {
             try
             {
-                var Class = await _classService.CreateClass(classDTO);
-                return Ok(Class);
+                var userPermissions = User.FindAll("Permission").Select(c => c.Value).ToList();
+                if (userPermissions is null)
+                {
+                    return Forbid();
+                }
+                else if (userPermissions.Contains("Thêm xóa sửa lớp"))
+                {
+                    var Class = await _classService.CreateClass(classDTO);
+                    return Ok();
+                }
+                else
+                {
+                    return Forbid("khong co quyen truy cap");
+                }
             }
             catch (Exception ex)
             {
@@ -29,12 +43,24 @@ namespace Course_Signup_System.Controllers
             }
         }
         [HttpGet]
-        public async Task<IActionResult> GetAllClass([FromQuery]int page, [FromQuery] int pagesize =10) 
+        public async Task<IActionResult> GetAllClass([FromQuery]int page = 1, [FromQuery] int pagesize =10) 
         {
             try
             {
-                var Classes = await _classService.GetAllClass(page, pagesize);
-                return Ok(Classes);
+                var userPermissions = User.FindAll("Permission").Select(c => c.Value).ToList();
+                if (userPermissions is null)
+                {
+                    return Forbid();
+                }
+                else if (userPermissions.Contains("Xem danh sách lớp"))
+                {
+                    var Classes = await _classService.GetAllClass(page, pagesize);
+                    return Ok(Classes);
+                }
+                else
+                {
+                    return Forbid("khong co quyen truy cap");
+                }
             }
             catch (Exception ex)
             {
@@ -46,8 +72,20 @@ namespace Course_Signup_System.Controllers
         {
             try
             {
-                var Class = await _classService.GetClassById(Id);
-                return Ok(Class);
+                var userPermissions = User.FindAll("Permission").Select(c => c.Value).ToList();
+                if (userPermissions is null)
+                {
+                    return Forbid();
+                }
+                else if (userPermissions.Contains("Xem danh sách lớp"))
+                {
+                    var Class = await _classService.GetClassById(Id);
+                    return Ok(Class);
+                }
+                else
+                {
+                    return Forbid("khong co quyen truy cap");
+                }
             }
             catch (Exception ex)
             {
@@ -59,8 +97,20 @@ namespace Course_Signup_System.Controllers
         {
             try
             {
-                var Class = await _classService.DeleteClass(Id);
-                return Ok(Class);
+                var userPermissions = User.FindAll("Permission").Select(c => c.Value).ToList();
+                if (userPermissions is null)
+                {
+                    return Forbid();
+                }
+                else if (userPermissions.Contains("Thêm xóa sửa lớp"))
+                {
+                    var Class = await _classService.DeleteClass(Id);
+                    return Ok(Class);
+                }
+                else
+                {
+                    return Forbid("khong co quyen truy cap");
+                }
             }
             catch (Exception ex)
             {
@@ -73,8 +123,21 @@ namespace Course_Signup_System.Controllers
             if (Id != classDTO.ClassId) return BadRequest("Id and classid incorrect");
             try
             {
-                var Class = await _classService.DeleteClass(Id);
-                return Ok(Class);
+                var userPermissions = User.FindAll("Permission").Select(c => c.Value).ToList();
+                if (userPermissions is null)
+                {
+                    return Forbid();
+                }
+                else if (userPermissions.Contains("Thêm xóa sửa lớp"))
+                {
+                    var Class = await _classService.DeleteClass(Id);
+                    return Ok(Class);
+                }
+                else
+                {
+                    return Forbid("khong co quyen truy cap");
+                }
+
             }
             catch (Exception ex)
             {
