+++ Course-Signup-System/Controllers/SchoolHolidayScheduleController.cs
@@ -1,6 +1,7 @@
 ﻿using Course_Signup_System.DTO;
 using Course_Signup_System.Entities;
 using Course_Signup_System.Services;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 
@@ -8,6 +9,7 @@ namespace Course_Signup_System.Controllers
 {
     [Route("api/[controller]")]
     [ApiController]
+    [Authorize]
     public class SchoolHolidayScheduleController : ControllerBase
     {
         private readonly ISchoolHolidayScheduleService _schoolHolidayScheduleService;
@@ -20,8 +22,17 @@ namespace Course_Signup_System.Controllers
         {
             try
             {
-                var sc = await _schoolHolidayScheduleService.CreateSchoolHolidaySchedule(schoolHolidayScheduleDTO);
-                return Ok(sc);
+                var userPermissions = User.FindAll("Permission").Select(c => c.Value).ToList();
+                if (userPermissions is null)
+                {
+                    return Forbid();
+                }
+                else if (userPermissions.Contains("Thêm xóa sửa lịch nghỉ"))
+                {
+                    var sc = await _schoolHolidayScheduleService.CreateSchoolHolidaySchedule(schoolHolidayScheduleDTO);
+                    return Ok(sc);
+                }
+                return Forbid();
             }
             catch (Exception ex)
             {
@@ -33,8 +44,17 @@ namespace Course_Signup_System.Controllers
         {
             try
             {
-                var sc = await _schoolHolidayScheduleService.DeleteSchoolHolidaySchedule(Id);
-                return Ok(sc);
+                var userPermissions = User.FindAll("Permission").Select(c => c.Value).ToList();
+                if (userPermissions is null)
+                {
+                    return Forbid();
+                }
+                else if (userPermissions.Contains("Thêm xóa sửa lịch nghỉ"))
+                {
+                    var sc = await _schoolHolidayScheduleService.DeleteSchoolHolidaySchedule(Id);
+                    return Ok(sc);
+                }
+                return Forbid();
                 
             }
             catch (Exception ex)
@@ -47,9 +67,17 @@ namespace Course_Signup_System.Controllers
         {
             try
             {
-                var sc = await _schoolHolidayScheduleService.GetAllSchoolHolidaySchedules();
-                return Ok(sc);
-                
+                var userPermissions = User.FindAll("Permission").Select(c => c.Value).ToList();
+                if (userPermissions is null)
+                {
+                    return Forbid();
+                }
+                else if (userPermissions.Contains("Danh sách lịch nghỉ"))
+                {
+                    var sc = await _schoolHolidayScheduleService.GetAllSchoolHolidaySchedules();
+                    return Ok(sc);
+                }
+                return Forbid();
             }
             catch (Exception ex)
             {
@@ -61,8 +89,17 @@ namespace Course_Signup_System.Controllers
         {
             try
             {
-                var sc = await _schoolHolidayScheduleService.GetSchoolHolidaySchedule(Id);
-                return Ok(sc);
+                var userPermissions = User.FindAll("Permission").Select(c => c.Value).ToList();
+                if (userPermissions is null)
+                {
+                    return Forbid();
+                }
+                else if (userPermissions.Contains("Danh sách lịch nghỉ"))
+                {
+                    var sc = await _schoolHolidayScheduleService.GetSchoolHolidaySchedule(Id);
+                    return Ok(sc);
+                }
+                return Forbid();
                 
             }
             catch (Exception ex)
@@ -75,8 +112,17 @@ namespace Course_Signup_System.Controllers
         {
             try
             {
-                var sc = await _schoolHolidayScheduleService.UpdateSchoolHolidaySchedule(schoolHolidayScheduleDTO);
-                return Ok(sc);
+                var userPermissions = User.FindAll("Permission").Select(c => c.Value).ToList();
+                if (userPermissions is null)
+                {
+                    return Forbid();
+                }
+                else if (userPermissions.Contains("Thêm xóa sửa lịch nghỉ"))
+                {
+                    var sc = await _schoolHolidayScheduleService.UpdateSchoolHolidaySchedule(schoolHolidayScheduleDTO);
+                    return Ok(sc);
+                }
+                return Forbid();
                 
             }
             catch (Exception ex)
