+++ Course-Signup-System/Controllers/StudentClassController.cs
@@ -1,12 +1,15 @@
 ﻿using Course_Signup_System.DTO;
 using Course_Signup_System.Services;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
+using System.Runtime.CompilerServices;
 
 namespace Course_Signup_System.Controllers
 {
     [Route("api/[controller]")]
     [ApiController]
+    [Authorize]
     public class StudentClassController : ControllerBase
     {
         private readonly IStudentClassService _studentClass;
@@ -14,6 +17,7 @@ namespace Course_Signup_System.Controllers
         {
             _studentClass = studentClass;
         }
+        
         [HttpPost]
         public async Task<IActionResult> CreateSudentClass([FromBody] StudentClassDTO studentClassDto)
         {
@@ -32,8 +36,20 @@ namespace Course_Signup_System.Controllers
         {
             try
             {
-                var result = await _studentClass.GetStudentClasses(page, pagesize);
-                return Ok(result);
+                var userPermissions = User.FindAll("Permission").Select(c => c.Value).ToList();
+                if (userPermissions is null)
+                {
+                    return Forbid();
+                }
+                else if (userPermissions.Contains("Danh sách học sinh trong lớp"))
+                {
+                    var result = await _studentClass.GetStudentClasses(page, pagesize);
+                    return Ok(result);
+                }
+                else
+                {
+                    return Forbid("khong co quyen truy cap");
+                }
             }
             catch (Exception ex)
             {
@@ -45,8 +61,20 @@ namespace Course_Signup_System.Controllers
         {
             try
             {
-                var result = await _studentClass.GetStudentClassById(Id);
-                return Ok(result);
+                var userPermissions = User.FindAll("Permission").Select(c => c.Value).ToList();
+                if (userPermissions is null)
+                {
+                    return Forbid();
+                }
+                else if (userPermissions.Contains("Danh sách học sinh trong lớp"))
+                {
+                    var result = await _studentClass.GetStudentClassById(Id);
+                    return Ok(result);
+                }
+                else
+                {
+                    return Forbid("khong co quyen truy cap");
+                }
             }
             catch (Exception ex)
             {
@@ -58,34 +86,70 @@ namespace Course_Signup_System.Controllers
         {
             try
             {
-                var result = await _studentClass.DeleteStudentClass(Id);
-                return Ok(result);
+                var userPermissions = User.FindAll("Permission").Select(c => c.Value).ToList();
+                if (userPermissions is null)
+                {
+                    return Forbid();
+                }
+                else if (userPermissions.Contains("Hủy đăng kí môn học"))
+                {
+                    var result = await _studentClass.DeleteStudentClass(Id);
+                    return Ok(result);
+                }
+                else
+                {
+                    return Forbid("Khong co quyen truy cap");
+                }
             }
             catch (Exception ex)
             {
                 return BadRequest(ex);
             }
         }
-        [HttpPut("{Id}")]
-        public async Task<IActionResult> UpdateStudentClass (int Id,StudentClassDTO studentClassDto)
-        {
-            try
-            {
-                var result = await _studentClass.UpdateStudentClass(Id, studentClassDto);
-                return Ok(result);
-            }
-            catch (Exception ex)
-            {
-                return BadRequest(ex);
-            }
-        }
+        //[HttpPut("{Id}")]
+        //public async Task<IActionResult> UpdateStudentClass(int Id, StudentClassDTO studentClassDto)
+        //{
+        //    try
+        //    {
+        //        var userPermissions = User.FindAll("Permission").Select(c => c.Value).ToList();
+        //        if (userPermissions is null)
+        //        {
+        //            return Forbid();
+        //        }
+        //        else if (userPermissions.Contains("Danh sách học sinh trong lớp"))
+        //        {
+        //            var result = await _studentClass.UpdateStudentClass(Id, studentClassDto);
+        //            return Ok(result);
+        //        }
+        //        else
+        //        {
+        //            return Forbid("khong co quyen truy cap");
+        //        }
+        //    }
+        //    catch (Exception ex)
+        //    {
+        //        return BadRequest(ex);
+        //    }
+        //}
         [HttpGet("student-Satus")]
         public async Task<IActionResult> GetStudentByStatus(bool status)
         {
             try
             {
-                var students = await _studentClass.GetStudentByStatus(status);
-                return Ok(students);
+                var userPermissions = User.FindAll("Permission").Select(c => c.Value).ToList();
+                if (userPermissions is null)
+                {
+                    return Forbid();
+                }
+                else if (userPermissions.Contains("Danh sách học sinh trong lớp"))
+                {
+                    var students = await _studentClass.GetStudentByStatus(status);
+                    return Ok(students);
+                }
+                else
+                {
+                    return Forbid("khong co quyen truy cap");
+                }
             }
             catch (Exception ex)
             {
