+++ Course-Signup-System/Controllers/TeachScheduleController.cs
@@ -1,5 +1,6 @@
 ﻿using Course_Signup_System.DTO;
 using Course_Signup_System.Services;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
@@ -8,6 +9,8 @@ namespace Course_Signup_System.Controllers
 {
     [Route("api/[controller]")]
     [ApiController]
+    [Authorize]
+
     public class TeachScheduleController : ControllerBase
     {
         private readonly ITeacherScheduleService _service;
@@ -20,8 +23,20 @@ namespace Course_Signup_System.Controllers
         {
             try
             {
-                var teachSchedule = await _service.CreateTeacherSchedule(teacherScheduleDTO);
-                return Ok(teachSchedule);
+                var userPermissions = User.FindAll("Permission").Select(c => c.Value).ToList();
+                if (userPermissions is null)
+                {
+                    return Forbid();
+                }
+                else if (userPermissions.Contains("Thêm xóa sửa lịch giảng dạy"))
+                {
+                    var teachSchedule = await _service.CreateTeacherSchedule(teacherScheduleDTO);
+                    return Ok(teachSchedule);
+                }
+                else
+                {
+                    return Forbid("khong co quyen");
+                }
             }
             catch (Exception ex)
             {
@@ -33,8 +48,20 @@ namespace Course_Signup_System.Controllers
         {
             try
             {
-                var teachSchedule = await _service.GetTeacherSchedule(page, pagesize);
-                return Ok(teachSchedule);
+                var userPermissions = User.FindAll("Permission").Select(c => c.Value).ToList();
+                if (userPermissions is null)
+                {
+                    return Forbid();
+                }
+                else if (userPermissions.Contains("Xem lịch giảng dạy"))
+                {
+                    var teachSchedule = await _service.GetTeacherSchedule(page, pagesize);
+                    return Ok(teachSchedule);
+                }
+                else
+                {
+                    return Forbid("không có quyền");
+                }
             }
             catch (Exception ex)
             {
@@ -46,8 +73,20 @@ namespace Course_Signup_System.Controllers
         {
             try
             {
-                var teachSchedule = await _service.GetTeacherScheduleById(Id);
-                return Ok(teachSchedule);
+                var userPermissions = User.FindAll("Permission").Select(c => c.Value).ToList();
+                if (userPermissions is null)
+                {
+                    return Forbid();
+                }
+                else if (userPermissions.Contains("Xem lịch giảng dạy"))
+                {
+                    var teachSchedule = await _service.GetTeacherScheduleById(Id);
+                    return Ok(teachSchedule);
+                }
+                else
+                {
+                    return Forbid("không có quyền");
+                }
             }
             catch (Exception ex)
             {
@@ -59,8 +98,20 @@ namespace Course_Signup_System.Controllers
         {
             try
             {
-                var teachSchedule = await _service.DeleteTeacherSchedule(Id);
-                return Ok(teachSchedule);
+                var userPermissions = User.FindAll("Permission").Select(c => c.Value).ToList();
+                if (userPermissions is null)
+                {
+                    return Forbid();
+                }
+                else if (userPermissions.Contains("Thêm xóa sửa lịch giảng dạy"))
+                {
+                    var teachSchedule = await _service.DeleteTeacherSchedule(Id);
+                    return Ok(teachSchedule);
+                }
+                else
+                {
+                    return Forbid("không có quyền truy cập");
+                }
             }
             catch (Exception ex)
             {
@@ -72,8 +123,20 @@ namespace Course_Signup_System.Controllers
         {
             try
             {
-                var teachSchedule = await _service.UpdateTeacherSchedule(Id, teacherScheduleDTO);
-                return Ok(teachSchedule);
+                var userPermissions = User.FindAll("Permission").Select(c => c.Value).ToList();
+                if (userPermissions is null)
+                {
+                    return Forbid();
+                }
+                else if (userPermissions.Contains("Thêm xóa sửa lịch giảng dạy"))
+                {
+                    var teachSchedule = await _service.UpdateTeacherSchedule(Id, teacherScheduleDTO);
+                    return Ok(teachSchedule);
+                }
+                else
+                {
+                    return Forbid("không có quyền");
+                }
             }
             catch (Exception ex)
             {
